Move tdr style to StdProcess.py
